shrink database by using uint unstead of ulong add more keyboard commands simplify string match functions fix nullpointer when adding titles add printf() formatting to popUp()

+++ database.h
@@ -15,7 +15,7 @@ struct dbentry_t {
 	char title[NAMELEN];		// Title info (from mp3)
 	char album[NAMELEN];		// Album info (from mp3)
 	char genre[NAMELEN];		// Album info (from mp3)
-	unsigned long played;		// play counter
+	unsigned int played;		// play counter
 };
 
 #define DBESIZE sizeof(struct dbentry_t)
